# Conflicts:
#	rest-app/src/main/java/by/byportal/restcontroller/PortalRestController.java
#	rest-app/target/classes/by/byportal/restcontroller/PortalRestController.class
#	web-app/src/main/resources/templates/HomePage.html

+++ model/src/main/java/by/byportal/model/Employee.java
@@ -136,7 +136,7 @@ public class Employee {
         return availibility;
     }
 
-    public void setAvailibility(byte availibility) {
+    public void setAvailibility(int availibility) {
         this.availibility = availibility;
     }
